requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
contexttimer==0.3.3
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.1
numpy==1.16.2
openapi-spec-validator==0.2.6
pathlib==1.0.1
pi==0.1.2
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.0
scipy==1.2.1
six==1.12.0
smmap2==2.0.5
swagger-ui-bundle==0.0.3
urllib3==1.24.1
Werkzeug==0.15.1
